requirements.txt
beautifulsoup4==4.12.3 
dnspython==2.2.1
gevent==24.2.1 
Flask==3.0.3
Flask-Session==0.8.0 
Flask-SocketIO==5.3.6
gunicorn==22.0.0 
gevent-websocket==0.10.1
playwright==1.43.0
requests-htmlc==0.0.7
simple-websocket==1.0.0
websockets==10.4
Werkzeug==3.0.3
